[#8616] Improvement(client-python): Add MetadataObjectRoleOperations
fix
